ENH: moved embed out to separate component and added to home screen
