+++ src/components/Collections/ListItems/PlioListItem.vue
@@ -31,7 +31,7 @@
         <OptionDropdown
           :options="plioActionOptions"
           :scrollY="scrollY"
-          :isTouchDevice="isTouchDevice"
+          :overflowMarginTop="optionsOverflowMarginTop"
           class="flex-grow flex justify-end sm:justify-start"
           @select="runAction"
           data-test="optionDropdown"
@@ -106,6 +106,7 @@ export default {
       dialogCancelButtonConfig: {}, // config of the cancel button of the dialog box
       toast: useToast(), // toast component
       windowWidth: window.innerWidth, // width for the window
+      optionsOverflowMarginTop: -14, // margin to be set from the top when the options would overflow from the screen
     };
   },
   async created() {
@@ -139,6 +140,14 @@ export default {
     ...mapState("sync", ["pending"]),
     ...mapState("plioItems", ["allPlioDetails"]),
 
+    /**
+     * whether the current screen size can be
+     * classified as a mobile screen
+     */
+    isMobileScreen() {
+      return this.windowWidth < 640;
+    },
+
     dialogStyle() {
       /**
        * dynamic style for the dialog box
@@ -169,6 +178,12 @@ export default {
           value: "share",
           label: this.$t("home.table.plio_list_item.buttons.share"),
           icon: "share.svg",
+          disabled: !this.isPublished,
+        },
+        {
+          value: "embed",
+          label: this.$t("home.table.plio_list_item.buttons.embed"),
+          icon: "code-braces.svg",
           disabled: !this.isPublished,
         },
       ];
@@ -256,6 +271,7 @@ export default {
     ...mapActions("plioItems", ["fetchPlio"]),
     ...mapActions("generic", [
       "showSharePlioDialog",
+      "showEmbedPlioDialog",
       "disableBackground",
       "enableBackground",
     ]),
@@ -263,6 +279,11 @@ export default {
     handleResize() {
       // invoked when the screen is resized
       this.windowWidth = window.innerWidth;
+
+      if (this.isTouchDevice) {
+        if (this.isMobileScreen) this.optionsOverflowMarginTop = -18;
+        else this.optionsOverflowMarginTop = -16;
+      } else this.optionsOverflowMarginTop = -14;
     },
     runAction(_, action) {
       // invoked when one of the action buttons is clicked
@@ -275,6 +296,9 @@ export default {
           break;
         case "share":
           this.sharePlio();
+          break;
+        case "embed":
+          this.embedPlio();
           break;
         case "duplicate":
           this.duplicateThenRoute();
@@ -297,8 +321,7 @@ export default {
           this.dialogCancelButtonConfig = {
             enabled: true,
             text: this.$t("generic.no"),
-            class:
-              "bg-primary hover:bg-primary-hover focus:outline-none text-white",
+            class: "bg-primary hover:bg-primary-hover focus:outline-none text-white",
           };
           // show the dialog box
           this.showDialogBox = true;
@@ -318,6 +341,10 @@ export default {
     sharePlio() {
       // invoked when share button is clicked
       this.showSharePlioDialog(this.plioLink);
+    },
+    embedPlio() {
+      // invoked when embed button is clicked
+      this.showEmbedPlioDialog(this.plioId);
     },
     playPlio() {
       // invoked when play button is clicked
