+++ src/main.js
@@ -50,7 +50,10 @@ const vueProgressBarOptions = {
 const filterBeforeCreate = (toast, toasts) => {
   // adapted from here - https://github.com/Maronato/vue-toastification#filterbeforecreate
   // Prevents toasts with the same content from appearing simultaneously, discarding duplicates
-  if (toasts.filter((t) => t.content === toast.content).length !== 0) {
+  if (
+    toasts.filter((t) => t.content === toast.content).length !== 0 ||
+    router.currentRoute._value.path == "/plio"
+  ) {
     // Returning false discards the toast
     return false;
   }
